refactor(kernel): Phase 1 clean up — remove relocated files, update imports, eliminate protocol wrappers
Remove old net/socket/ and net/protocol/ files that were relocated to net/smoltcp/;
update all import paths and module declarations to reflect the new structure;
eliminate TcpProtocol/UdpProtocol/IcmpProtocol wrapper types in favor of factory structs
in protocol/mod.rs; restore net_manager/loopback/virtio to original locations.
